refactors completetask conversation
