+++ elsje.js
@@ -5,6 +5,7 @@ var api = require('./svlo-api')
 var post = require('./post')
 var newtask = require('./conversations/newtask')
 var instant = require('./conversations/instant')
+var completetask = require('./conversations/completetask')
 
 var Botkit = require('botkit')
 var Colormap = require('colormap')
@@ -29,11 +30,11 @@ var bot = controller.spawn({
 var reconnectcounter = 0
 controller.on('rtm_open', post.reconnect(bot, reconnectcounter, debug, function (increasedCounter) {
   reconnectcounter = increasedCounter
-})
+}))
 
-controller.hears(['shutdown'], 'direct_message,direct_mention,mention', functions.shutdown(bot, message))
+controller.hears(['shutdown'], 'direct_message,direct_mention,mention', functions.shutdown)
 
-controller.hears(['ken ik jou', 'wie ben jij', 'hoe lang ben je al wakker', 'uptime', 'identify yourself', 'who are you', 'what is your name'], 'direct_message,direct_mention,mention', functions.uptime(bot, message))
+controller.hears(['ken ik jou', 'wie ben jij', 'hoe lang ben je al wakker', 'uptime', 'identify yourself', 'who are you', 'what is your name'], 'direct_message,direct_mention,mention', functions.uptime)
 
 controller.hears(['nieuwe taak', 'voeg toe', 'taak (.*)voegen'], 'direct_mention,mention,direct_message', function (bot, message) {
   bot.createConversation(message, newtask.conversation)
@@ -41,53 +42,9 @@ controller.hears(['nieuwe taak', 'voeg toe', 'taak (.*)voegen'], 'direct_mention
 
 controller.hears(['instanttaak (.*)'], 'direct_message', instant.taak)
 
-
-
 controller.hears(['taak (.*)afronden', 'taak (.*)afvinken', 'ik ben klaar', 'taak (.*)gedaan'], 'direct_mention,mention,direct_message', function (bot, message) {
-  bot.startConversation(message, completeTask)
+  bot.createConversation(message, completetask.conversation)
 })
-
-var completeTask = function (response, convo) {
-  if (!isNaN(parseInt(convo.source_message.match[1], 10))) {
-    finishtask(convo, parseInt(convo.source_message.match[1], 10))
-  } else {
-    showListGetNum(response, convo, TaskDone)
-  }
-}
-
-var TaskDone = function (response, convo) {
-  convo.on('end', function (convo) {
-    if (convo.status === 'completed') {
-      var res = convo.extractResponses()
-      var number = parseInt(res['Kan je mij het nummer geven van de taak die van de lijst af mag?'], 10)
-      finishtask(convo, number)
-    }
-  })
-}
-
-var finishtask = function (convo, taskNumber) {
-  var userId = convo.source_message.user
-  api.showSingleTask(taskNumber, function (err, task) {
-    if (err) {
-      functions.postMessage(bot, 'Er is iets misggegaan bij het weergeven van de taak', convo.source_message.channel)
-    } else {
-      api.completeTask(taskNumber, function (error) {
-        if (error) {
-          functions.postMessage(bot, 'Er is iets misgegaan bij het bijwerken van de taak.', convo.source_message.channel)
-        } else {
-          task.status = 1
-          var message = {
-            'fallback': 'Taak van <@' + task.responsibleid + '> door <@' + userId + '> afgerond: ' + task.task,
-            'color': 'good',
-            'pretext': 'Taak afgerond door <@' + userId + '>.'
-          }
-          functions.postSingleTask(bot, task, message)
-        }
-      })
-    }
-  })
-  convo.stop()
-}
 
 controller.hears(['update deadline', 'deadline veranderen', 'andere deadline'], 'direct_mention,mention,direct_message', function (bot, message) {
   bot.startConversation(message, DeadlineNumber)
